workspace/first_mcp_project/pyproject.toml
[project]
name = "first-mcp-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "anthropic>=0.74.0",
    "arxiv>=2.3.1",
    "httpx>=0.28.1",
    "mcp[cli]>=1.21.2",
    "nest-asyncio>=1.6.0",
    "openai>=2.8.1",
    "python-dotenv>=1.2.1",
]
